refactor StrategyPerformanceChart part 2:
extract PerformanceChart inner/child component
