+++ src/lib/chart/index.ts
@@ -8,4 +8,5 @@ export { default as HudRow } from './HudRow.svelte';
 export { default as HudMetric } from './HudMetric.svelte';
 export { default as Marker } from './Marker.svelte';
 export { default as PairCandleChart } from './PairCandleChart.svelte';
+export { default as PerformanceChart } from './PerformanceChart.svelte';
 export { default as ReserveInterestChart } from './ReserveInterestChart.svelte';
